myfile: proper 64-bit support

+++ rtengine/dcraw.cc
@@ -9518,7 +9518,7 @@ void CLASS identify()
       "Nikon", "Nokia", "Olympus", "Ricoh", "Pentax", "Phase One",
       "Samsung", "Sigma", "Sinar", "Sony", "YI" };
   char head[32], *cp;
-  int hlen, flen, fsize, zero_fsize=1, i, c;
+  ssize_t hlen, flen, fsize, zero_fsize=1, i, c;
   struct jhead jh;
 
   tiff_flip = flip = filters = UINT_MAX;	/* unknown */
